+++ ci/release_notes.md
@@ -1,1 +1,3 @@
-Empty - please add release notes here
+## Bug Fixes
+- trades query should always denote volume in BTC
+- return buy and sell fields in ticker query